refactor: rename userProfile.jsx (hook) to useUserProfile.jsx
